# src/main/java/com/leonardgarden/greenhouse/equipmentservice/api/HistoryApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.leonardgarden.greenhouse.equipmentservice.api;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.leonardgarden.greenhouse.equipmentservice.model.EquipmentHistoryLog;
import io.swagger.annotations.*;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import java.util.Optional;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-04-15T16:33:30.590-05:00")

@Api(value = "history", description = "the history API")
public interface HistoryApi {

    Logger log = LoggerFactory.getLogger(HistoryApi.class);

    default Optional<ObjectMapper> getObjectMapper() {
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest() {
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @ApiOperation(value = "Adds a history record for an equipment object", nickname = "historyEquipmentPost", notes = "", tags={ "equipment history", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "History data added successfully"),
        @ApiResponse(code = 500, message = "Server Error") })
    @RequestMapping(value = "/history/equipment/",
        method = RequestMethod.POST)
    default ResponseEntity<Void> historyEquipmentPost(@ApiParam(value = "JSON data representing an equipment history object", required = true) @Valid @RequestBody EquipmentHistoryLog equipmentHistoryData) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default HistoryApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}